rename items controllers
