+++ controllers/web/item-detail.js
