+++ index.js
@@ -63,11 +63,11 @@ const createFriendApiController = require('./controllers/api/createFriendApi')
 const deleteFriendApiController = require('./controllers/api/deleteFriendApi')
 
 const indexController = require('./controllers/web/index-controller')
-const upcomingItemsController = require('./controllers/web/upcomingItems')
-const itemDetailController = require('./controllers/web/itemDetail')
-const storeItemsController = require('./controllers/web/storeItems')
-const itemsListController = require('./controllers/web/itemList')
-const itemsSaveController = require('./controllers/web/saveItems')
+const upcomingItemsController = require('./controllers/web/item-upcoming')
+const itemDetailController = require('./controllers/web/item-detail')
+const storeItemsController = require('./controllers/web/item-store')
+const itemsListController = require('./controllers/web/item-list')
+const itemsSaveController = require('./controllers/web/item-list-save')
 const getUserStatsController = require('./controllers/web/get-user-stats')
 const getUserStatsSubmitController = require('./controllers/web/get-user-stats-submit')
 const getChartsController = require('./controllers/web/get-charts-controller')
